app/config/database.py
import os
from dotenv import load_dotenv

# Carga variables de entorno de .env
load_dotenv()

class DatabaseConfig:
    SQLALCHEMY_DATABASE_URI = os.getenv(
        "DATABASE_URL",
        "sqlite:///./products.db"
    )
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    SECRET_KEY = os.getenv("SECRET_KEY", "fallback-secret")
