requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
flasgger==0.9.7.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.4.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
rpds-py==0.25.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==2.3.7
